% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kernelMatrix_cpp}
\alias{kernelMatrix_cpp}
\title{Compute the kernel matrix between two sets of data points.}
\usage{
kernelMatrix_cpp(X1, X2, kernel_spec)
}
\arguments{
\item{X1}{First set of data points (each row is a data point)}

\item{X2}{Second set of data points (each row is a data point)}

\item{kernel_spec}{R list describing the kernel}
}
\value{
Kernel matrix K
}
\description{
Compute the kernel matrix between two sets of data points.
}
