% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kernelScalar_cpp}
\alias{kernelScalar_cpp}
\title{Evaluate the kernel function for a single pair of data points.}
\usage{
kernelScalar_cpp(x1, x2, kernel_spec)
}
\arguments{
\item{x1}{First data point (vector)}

\item{x2}{Second data point (vector)}

\item{kernel_spec}{R list describing the kernel}
}
\value{
Kernel value k(x1, x2)
}
\description{
Evaluate the kernel function for a single pair of data points.
}
