% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{print.kernel_spec}
\alias{print.kernel_spec}
\title{Print method for kernel specifications}
\usage{
\method{print}{kernel_spec}(x, ...)
}
\arguments{
\item{x}{Kernel specification}

\item{...}{Additional arguments (not used)}
}
\value{
Invisible NULL
}
\description{
Print method for kernel specifications
}
