% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{print_kernel}
\alias{print_kernel}
\title{Print a kernel specification}
\usage{
print_kernel(kernel, indent = 0)
}
\arguments{
\item{kernel}{Kernel specification}

\item{indent}{Indentation level (for nested printing)}
}
\value{
Invisible NULL
}
\description{
Internal function to recursively print kernel specifications.
}
\keyword{internal}
